+++ src/Twig/RouteExtension.php
@@ -1,0 +1,33 @@
+<?php
+
+// src/Twig/AppExtension.php
+namespace App\Twig;
+
+use Symfony\Component\Routing\RouteCollection;
+use Symfony\Component\Routing\Router;
+use Symfony\Component\Routing\RouterInterface;
+use Twig\Extension\AbstractExtension;
+use Twig\TwigFilter;
+use Twig\TwigFunction;
+
+class RouteExtension extends AbstractExtension
+{
+    private Router $router;
+
+    public function __construct(RouterInterface $router)
+    {
+        $this->router = $router;
+    }
+
+    public function getFunctions()
+    {
+        return [
+            new TwigFunction('route_exists', [$this, 'routeExists']),
+        ];
+    }
+
+    public function routeExists(string $routeName)
+    {
+        return null !== $this->router->getRouteCollection()->get($routeName);
+    }
+}