+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 * ADD public - liste des serveurs DCS de l'association (perun)
 * ADD public - liste des joueurs connectés aux serveurs DCS (perun)
 * ADD public - page "le bureau" gérée en dynamique depuis les statuts des membres du bureau
+* ADD public - affichage des pages dynamiques (CMS)
 * ADD roster - affichage des pilotes par groupe et par module (dynamique)
 * ADD admin - interface d'admnistration
 * ADD admin - administration des joueurs DCS (perun)
 * ADD admin - administration des utilisateurs du site
+* ADD admin - interface d'admnistration des pages (CMS)
 * ADD utilisateur - inscription (+ recaptcha)
 * ADD utilisateur - mot de passe perdu (+ recaptcha), validation par email
 * ADD utilisateur - interface de connexion / déconnexion